+++ tracker/tracker/src/webworker/MessageEncoder.gen.ts
@@ -254,8 +254,12 @@ export default class MessageEncoder extends PrimitiveEncoder {
       return  this.string(msg[1]) && this.string(msg[2])
     break
 
-    case Messages.Type.BatchMetadata:
+    case Messages.Type.BatchMetadataDeprecated:
       return  this.uint(msg[1]) && this.uint(msg[2]) && this.uint(msg[3]) && this.int(msg[4]) && this.string(msg[5])
+    break
+
+    case Messages.Type.BatchMessageOffsets:
+      return  true 
     break
 
     case Messages.Type.NetworkRequest:
@@ -268,6 +272,10 @@ export default class MessageEncoder extends PrimitiveEncoder {
 
     case Messages.Type.ResourceTiming:
       return  this.uint(msg[1]) && this.uint(msg[2]) && this.uint(msg[3]) && this.uint(msg[4]) && this.uint(msg[5]) && this.uint(msg[6]) && this.string(msg[7]) && this.string(msg[8]) && this.uint(msg[9]) && this.boolean(msg[10]) && this.uint(msg[11]) && this.uint(msg[12]) && this.uint(msg[13]) && this.uint(msg[14]) && this.uint(msg[15]) && this.uint(msg[16]) && this.uint(msg[17])
+    break
+
+    case Messages.Type.BatchMetadata:
+      return  this.uint(msg[1]) && this.uint(msg[2]) && this.uint(msg[3]) && this.int(msg[4]) && this.string(msg[5]) && this.int(msg[6]) && this.int(msg[7])
     break
 
     case Messages.Type.Incident: